whole docs now run new repl - move MicroRepl code to MiniRepl - fix a ssr bug
