+++ website/src/docs/MiniRepl.jsx
@@ -1,84 +1,159 @@
-import { evalScope, controls, noteToMidi } from '@strudel.cycles/core';
-import { initAudioOnFirstClick } from '@strudel.cycles/webaudio';
-import { useEffect, useState } from 'react';
-import { prebake } from '../repl/prebake';
-import { themes, settings } from '../repl/themes.mjs';
-import './MiniRepl.css';
-import { useSettings } from '../settings.mjs';
+import { useState, useRef, useCallback, useMemo, useEffect } from 'react';
+import { Icon } from './Icon';
+import { silence, getPunchcardPainter, noteToMidi } from '@strudel.cycles/core';
+import { transpiler } from '@strudel.cycles/transpiler';
+import { getAudioContext, webaudioOutput } from '@strudel.cycles/webaudio';
+import { StrudelMirror } from '@strudel/codemirror';
+// import { prebake } from '@strudel/repl';
+import { prebake } from '../repl/prebake.mjs';
+import { loadModules } from '../repl/util.mjs';
 import Claviature from '@components/Claviature';
 
-let modules;
+let prebaked, modulesLoading;
 if (typeof window !== 'undefined') {
-  modules = evalScope(
-    controls,
-    import('@strudel.cycles/core'),
-    import('@strudel.cycles/tonal'),
-    import('@strudel.cycles/mini'),
-    import('@strudel.cycles/midi'),
-    import('@strudel.cycles/xen'),
-    import('@strudel.cycles/webaudio'),
-    import('@strudel.cycles/osc'),
-    import('@strudel.cycles/csound'),
-    import('@strudel.cycles/soundfonts'),
-    import('@strudel/hydra'),
-  );
-}
-
-if (typeof window !== 'undefined') {
-  initAudioOnFirstClick();
-  prebake();
+  prebaked = prebake();
+  modulesLoading = loadModules();
 }
 
 export function MiniRepl({
-  tune,
-  drawTime,
+  tune: code,
+  hideHeader = false,
+  canvasHeight = 100,
+  onTrigger,
   punchcard,
   punchcardLabels = true,
-  span = [0, 4],
-  canvasHeight = 100,
-  hideHeader,
   claviature,
   claviatureLabels,
 }) {
-  const [Repl, setRepl] = useState();
-  const { theme, keybindings, fontSize, fontFamily, isLineNumbersDisplayed, isActiveLineHighlighted } = useSettings();
+  const id = useMemo(() => s4(), []);
+  const canvasId = useMemo(() => `canvas-${id}`, [id]);
+  const shouldDraw = !!punchcard || !!claviature;
+  const shouldShowCanvas = !!punchcard;
+  const drawTime = punchcard ? [0, 4] : [0, 0];
   const [activeNotes, setActiveNotes] = useState([]);
-  useEffect(() => {
-    // we have to load this package on the client
-    // because codemirror throws an error on the server
-    Promise.all([import('@strudel.cycles/react'), modules])
-      .then(([res]) => setRepl(() => res.MiniRepl))
-      .catch((err) => console.error(err));
-  }, []);
-  return Repl ? (
-    <div className="mb-4 mini-repl">
-      <Repl
-        tune={tune}
-        hideOutsideView={true}
-        drawTime={claviature ? [0, 0] : drawTime}
-        punchcard={punchcard}
-        punchcardLabels={punchcardLabels}
-        span={span}
-        canvasHeight={canvasHeight}
-        theme={themes[theme]}
-        hideHeader={hideHeader}
-        keybindings={keybindings}
-        fontFamily={fontFamily}
-        fontSize={fontSize}
-        isLineNumbersDisplayed={isLineNumbersDisplayed}
-        isActiveLineHighlighted={isActiveLineHighlighted}
-        onPaint={
-          claviature
-            ? (ctx, time, haps, drawTime) => {
-                const active = haps
-                  .map((hap) => hap.value.note)
-                  .filter(Boolean)
-                  .map((n) => (typeof n === 'string' ? noteToMidi(n) : n));
-                setActiveNotes(active);
-              }
-            : undefined
+
+  const init = useCallback(({ code, shouldDraw }) => {
+    const drawContext = shouldDraw ? document.querySelector('#' + canvasId)?.getContext('2d') : null;
+    let onDraw;
+    if (shouldDraw) {
+      onDraw = (haps, time, frame, painters) => {
+        painters.length && drawContext?.clearRect(0, 0, drawContext.canvas.width * 2, drawContext.canvas.height * 2);
+        painters?.forEach((painter) => {
+          // ctx time haps drawTime paintOptions
+          painter(drawContext, time, haps, drawTime, { clear: false });
+        });
+      };
+    }
+
+    const editor = new StrudelMirror({
+      id,
+      defaultOutput: webaudioOutput,
+      getTime: () => getAudioContext().currentTime,
+      transpiler,
+      autodraw: !!shouldDraw,
+      root: containerRef.current,
+      initialCode: '// LOADING',
+      pattern: silence,
+      drawTime,
+      onDraw,
+      editPattern: (pat, id) => {
+        if (onTrigger) {
+          pat = pat.onTrigger(onTrigger, false);
         }
-      />
+        if (claviature) {
+          editor?.painters.push((ctx, time, haps, drawTime) => {
+            const active = haps
+              .map((hap) => hap.value.note)
+              .filter(Boolean)
+              .map((n) => (typeof n === 'string' ? noteToMidi(n) : n));
+            setActiveNotes(active);
+          });
+        }
+        if (punchcard) {
+          editor?.painters.push(getPunchcardPainter({ labels: !!punchcardLabels }));
+        }
+        return pat;
+      },
+      prebake: async () => Promise.all([modulesLoading, prebaked]),
+      onUpdateState: (state) => {
+        setReplState({ ...state });
+      },
+    });
+    // init settings
+    editor.setCode(code);
+    editorRef.current = editor;
+  }, []);
+
+  const [replState, setReplState] = useState({});
+  const { started, isDirty, error } = replState;
+  const editorRef = useRef();
+  const containerRef = useRef();
+  const [client, setClient] = useState(false);
+  useEffect(() => {
+    setClient(true);
+    if (!editorRef.current) {
+      setTimeout(() => {
+        init({ code, shouldDraw });
+      });
+    }
+    return () => {
+      editorRef.current?.clear();
+    };
+  }, []);
+
+  if (!client) {
+    return <pre>{code}</pre>;
+  }
+
+  return (
+    <div className="overflow-hidden rounded-t-md bg-background border border-lineHighlight">
+      {!hideHeader && (
+        <div className="flex justify-between bg-lineHighlight">
+          <div className="flex">
+            <button
+              className={cx(
+                'cursor-pointer w-16 flex items-center justify-center p-1 border-r border-lineHighlight text-foreground bg-lineHighlight hover:bg-background',
+                started ? 'animate-pulse' : '',
+              )}
+              onClick={() => editorRef.current?.toggle()}
+            >
+              <Icon type={started ? 'stop' : 'play'} />
+            </button>
+            <button
+              className={cx(
+                'w-16 flex items-center justify-center p-1 text-foreground border-lineHighlight bg-lineHighlight',
+                isDirty ? 'text-foreground hover:bg-background cursor-pointer' : 'opacity-50 cursor-not-allowed',
+              )}
+              onClick={() => editorRef.current?.evaluate()}
+            >
+              <Icon type="refresh" />
+            </button>
+          </div>
+        </div>
+      )}
+      <div className="overflow-auto relative p-1">
+        <div ref={containerRef}></div>
+        {error && <div className="text-right p-1 text-md text-red-200">{error.message}</div>}
+      </div>
+      {shouldShowCanvas && (
+        <canvas
+          id={canvasId}
+          className="w-full pointer-events-none border-t border-lineHighlight"
+          height={canvasHeight}
+          ref={(el) => {
+            if (el && el.width !== el.clientWidth) {
+              el.width = el.clientWidth;
+            }
+          }}
+        ></canvas>
+      )}
+      {/* !!log.length && (
+      <div className="bg-gray-800 rounded-md p-2">
+        {log.map(({ message }, i) => (
+          <div key={i}>{message}</div>
+        ))}
+      </div>
+    ) */}
       {claviature && (
         <Claviature
           options={{
@@ -90,7 +165,16 @@ export function MiniRepl({
         />
       )}
     </div>
-  ) : (
-    <pre>{tune}</pre>
   );
+}
+
+function cx(...classes) {
+  // : Array<string | undefined>
+  return classes.filter(Boolean).join(' ');
+}
+
+function s4() {
+  return Math.floor((1 + Math.random()) * 0x10000)
+    .toString(16)
+    .substring(1);
 }
